[GH-#97] Scheduling work is done in a separate module again, checking for unknown blocks is done every minute
