+++ apps/aecore/lib/aecore/peers/worker/supervisor.ex
@@ -8,7 +8,8 @@ defmodule Aecore.Peers.Worker.Supervisor do
   def init(:ok) do
     children = [
       Aecore.Peers.Worker,
-      Aecore.Peers.Sync
+      Aecore.Peers.Sync,
+      Aecore.Peers.Scheduler
     ]
 
     Supervisor.init(children, strategy: :one_for_one)
